Align workshop routing with other activities
