+++ frontend/src/types/flashcards.ts
@@ -1,0 +1,4 @@
+export interface Flashcard {
+  question: string;
+  reponse: string;
+}
